+++ d2/src/main/java/com/linkedin/d2/xds/XdsClientImpl.java
@@ -222,10 +222,9 @@ public class XdsClientImpl extends XdsClient
       try
       {
         XdsD2.Node d2Node = resource.getResource().unpack(XdsD2.Node.class);
-        if (d2Node != null && d2Node.getData().isEmpty() ) {
+        if (d2Node != null && d2Node.getData().isEmpty()) {
           _log.warn("Received a Node response with no data, resource is : {}", resourceName);
-        }
-        else {
+        } else {
           updates.put(resourceName, new NodeUpdate(resource.getVersion(), d2Node));
         }
       } catch (InvalidProtocolBufferException e)
@@ -234,7 +233,7 @@ public class XdsClientImpl extends XdsClient
         errors.add("Failed to unpack Node response");
       }
     }
-    handleResponseACK(data.getResourceType(), data.getNonce(), errors);
+    sendAckOrNack(data.getResourceType(), data.getNonce(), errors);
     handleResourceUpdate(updates, data.getResourceType());
     handleResourceRemoval(data.getRemovedResources(), data.getResourceType());
   }
@@ -251,9 +250,9 @@ public class XdsClientImpl extends XdsClient
       {
         XdsD2.D2URIMap uriMap = resource.getResource().unpack(XdsD2.D2URIMap.class);
         Map<String, XdsD2.D2URI> nodeData = uriMap.getUrisMap();
-        if(nodeData.isEmpty()){
+        if (nodeData.isEmpty()) {
           _log.warn("Received a D2URIMap response with no data, resource is : {}", resourceName);
-        }else{
+        } else {
           updates.put(resourceName, new D2URIMapUpdate(resource.getVersion(), nodeData));
         }
       } catch (InvalidProtocolBufferException e)
@@ -262,11 +261,11 @@ public class XdsClientImpl extends XdsClient
         errors.add("Failed to unpack D2URIMap response");
       }
     }
-    handleResponseACK(data.getResourceType(), data.getNonce(), errors);
+    sendAckOrNack(data.getResourceType(), data.getNonce(), errors);
     handleResourceUpdate(updates, data.getResourceType());
   }
 
-  private void handleResponseACK(ResourceType type, String nonce, List<String> errors) {
+  private void sendAckOrNack(ResourceType type, String nonce, List<String> errors) {
     if (errors.isEmpty()) {
       _adsStream.sendAckRequest(type, nonce);
     } else {
@@ -275,29 +274,23 @@ public class XdsClientImpl extends XdsClient
     }
   }
 
-  private void handleResourceUpdate(Map<String, ? extends ResourceUpdate> updates, ResourceType type)
-  {
-
-
-    for (Map.Entry<String, ? extends ResourceUpdate> entry : updates.entrySet())
-    {
+  private void handleResourceUpdate(Map<String, ? extends ResourceUpdate> updates, ResourceType type) {
+    for (Map.Entry<String, ? extends ResourceUpdate> entry : updates.entrySet()) {
       String resourceName = entry.getKey();
       ResourceUpdate resourceUpdate = entry.getValue();
       ResourceSubscriber subscriber = getResourceSubscriberMap(type).get(resourceName);
-      if (subscriber != null)
-      {
+      if (subscriber != null) {
         subscriber.onData(resourceUpdate);
       }
     }
   }
 
-  private void handleResourceRemoval(List<String> removedResources, ResourceType type)
-  {
-    for (String resourceName : removedResources)
-    {
+  private void handleResourceRemoval(List<String> removedResources, ResourceType type) {
+    for (String resourceName : removedResources) {
+      _xdsClientJmx.incrementResourceNotFoundCount();
+      _log.info("Resource {} need to be removed", resourceName);
       ResourceSubscriber subscriber = getResourceSubscriberMap(type).get(resourceName);
-      if (subscriber != null)
-      {
+      if (subscriber != null) {
         subscriber.onRemoval();
       }
     }
@@ -390,7 +383,9 @@ public class XdsClientImpl extends XdsClient
         _log.debug("Received resource update data equal to the current data. Will not perform the update.");
         return;
       }
-      _data = data;
+      if (data != null) { // null value guard to avoid overwriting the property with null
+        _data = data;
+      }
       for (ResourceWatcher watcher : _watchers)
       {
         notifyWatcher(watcher, data);
@@ -413,9 +408,11 @@ public class XdsClientImpl extends XdsClient
       }
     }
 
-    private void onRemoval(){
-      for (ResourceWatcher watcher : _watchers)
-      {
+    private void onRemoval() {
+      // When the client receive the removal data from INDIS or ZK,
+      // client side doesn't delete the data from the cache which is just a design choice by now.
+      // So to avoid the eventbus watcher timeout, in there directly notify the watcher with cache data
+      for (ResourceWatcher watcher : _watchers) {
         notifyWatcher(watcher, _data);
       }
     }